backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
torch>=2.0.0
google-generativeai>=0.3.0
requests>=2.31.0
